Add extra checks around temp directory deletion

+++ command_line/commoncommandlinefunctions.cpp
@@ -977,5 +977,8 @@ bool createBlastTempDirectory()
 
 void deleteBlastTempDirectory()
 {
-    QDir(g_blastSearch->m_tempDirectory).removeRecursively();
+    if (g_blastSearch->m_tempDirectory != "" &&
+            QDir(g_blastSearch->m_tempDirectory).exists() &&
+            QDir(g_blastSearch->m_tempDirectory).dirName().contains("bandage_temp"))
+        QDir(g_blastSearch->m_tempDirectory).removeRecursively();
 }
